Minor bugfixes at deletion in logic.

+++ HKDXX6_HFT_2023241.Logic/Classes/CaseLogic.cs
@@ -82,6 +82,12 @@ namespace HKDXX6_HFT_2023241.Logic
 
         public void Delete(int ID)
         {
+            var c = CaseRepo.Read(ID);
+            if (c == null)
+            {
+                throw new ArgumentException("Case does not exist.");
+            }
+
             CaseRepo.Delete(ID);
         }
